+++ pmd-core/src/main/java/net/sourceforge/pmd/RuleSet.java
@@ -12,6 +12,7 @@ import java.util.Collections;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Objects;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -38,6 +39,8 @@ public class RuleSet implements ChecksumAware {
 
     private static final Logger LOG = Logger.getLogger(RuleSet.class.getName());
     private static final String MISSING_RULE = "Missing rule";
+    private static final String MISSING_RULESET_DESCRIPTION = "RuleSet description must not be null";
+    private static final String MISSING_RULESET_NAME = "RuleSet name must not be null";
 
     private final long checksum;
 
@@ -64,8 +67,8 @@ public class RuleSet implements ChecksumAware {
     private RuleSet(final RuleSetBuilder builder) {
         checksum = builder.checksum;
         fileName = builder.fileName;
-        name = builder.name;
-        description = builder.description;
+        name = Objects.requireNonNull(builder.name, MISSING_RULESET_NAME);
+        description = Objects.requireNonNull(builder.description, MISSING_RULESET_DESCRIPTION);
         // TODO: ideally, the rules would be unmodifiable, too. But removeDysfunctionalRules might change the rules.
         rules = builder.rules;
         excludePatterns = Collections.unmodifiableList(builder.excludePatterns);
@@ -76,8 +79,8 @@ public class RuleSet implements ChecksumAware {
     }
 
     /* package */ static class RuleSetBuilder {
-        public String description = "";
-        public String name = "";
+        public String description;
+        public String name;
         public String fileName;
         private final List<Rule> rules = new ArrayList<>();
         private final List<String> excludePatterns = new ArrayList<>(0);
@@ -347,13 +350,17 @@ public class RuleSet implements ChecksumAware {
         }
 
         public RuleSetBuilder withName(final String name) {
-            this.name = name;
+            this.name = Objects.requireNonNull(name, MISSING_RULESET_NAME);
             return this;
         }
 
         public RuleSetBuilder withDescription(final String description) {
-            this.description = description;
+            this.description = Objects.requireNonNull(description, MISSING_RULESET_DESCRIPTION);
             return this;
+        }
+
+        public boolean hasDescription() {
+            return this.description != null;
         }
 
         public String getName() {